requirements.txt
fastapi
uvicorn
websockets
opencv-python
paho-mqtt
jinja2
python-multipart
mediapipe
